Update format strings in docs/examples for correct signedness.
The latest version of Clang 19 detects sign mismatches in format
strings when -Wformat is enabled.

Reviewed-on: https://skia-review.googlesource.com/c/skia/+/842477
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>

+++ docs/examples/ImageInfo_minRowBytes64.cpp
@@ -11,7 +11,7 @@ void draw(SkCanvas* canvas) {
         uint64_t minRowBytes = imageInfo.minRowBytes64();
         bool widthTooLarge = (uint64_t) (int32_t) minRowBytes != minRowBytes;
         SkDebugf("RGBA_F16 width %d (0x%08x) %s\n",
-                width, width, widthTooLarge ? "too large" : "OK");
+                 width, (unsigned int)width, widthTooLarge ? "too large" : "OK");
     }
 }
 }  // END FIDDLE
